requirements.txt
pyautogen>=0.8.0
python-dotenv>=1.0.1
pytest>=8.2.0
